+++ R/dbi-driver.R
@@ -93,7 +93,7 @@ setMethod("show", "OdbcDriver", function(object) {
 #'   additional arguments will be appended to this string.
 #' @param bigint The R type that `SQL_BIGINT` types should be mapped to.
 #'   Default is [bit64::integer64], which allows the full range of 64 bit
-#'   integers.
+#'   integers. Arrow results (see [DBI-arrow]) always use `int64`.
 #' @param timeout Time in seconds to timeout the connection attempt. Setting a
 #'   timeout of `Inf` or `NA` indicates no timeout. Defaults to 10 seconds.
 #'